standardized data type for time variables and added function comments.

+++ death.c
@@ -6,13 +6,13 @@
 /*   By: etien <etien@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/11 10:36:06 by etien             #+#    #+#             */
-/*   Updated: 2024/09/17 11:28:47 by etien            ###   ########.fr       */
+/*   Updated: 2024/09/17 14:20:01 by etien            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-// This function is run by the death monitor detached thread.
+// This function is run by the death monitor thread.
 // If a philosopher has died, its corresponding death monitor will
 // set the shared dead_philo flag to true.
 // In the beginning, the thread will sleep until it is close to time_to_die.
@@ -25,7 +25,7 @@
 void	*check_philo_death(void *arg)
 {
 	t_philo		*philo;
-	long		time_to_die;
+	int			time_to_die;
 
 	philo = (t_philo *) arg;
 	time_to_die = philo->data->time_to_die;
@@ -47,10 +47,10 @@ void	*check_philo_death(void *arg)
 
 // This function will set the dead_philo boolean to true.
 // It is protected behind the death mutex.
-// printf is used here because print is set up to stop working
+// printf is used here because print is designed to stop working
 // once a philosopher has died.
-// The stop_simulation boolean is also set so that only the death
-// of ONE philosopher is printed.
+// The stop_simulation boolean is also set so that the death
+// of only ONE philosopher is printed.
 // Whenever multiple threads have access to the same shared
 // variable, getting and setting functions for that variable
 // have to be locked behind a mutex to make it thread-safe.
@@ -61,7 +61,7 @@ void	set_philo_dead(t_philo *philo)
 	if (!philo->data->stop_simulation)
 	{
 		pthread_mutex_lock(&philo->data->print_mutex);
-		printf("%ld %i %s", timestamp() - philo->data->start_time,
+		printf("%lld %i %s", timestamp() - philo->data->start_time,
 			philo->id, DIED);
 		pthread_mutex_unlock(&philo->data->print_mutex);
 		philo->data->stop_simulation = true;
@@ -72,6 +72,8 @@ void	set_philo_dead(t_philo *philo)
 // This function will check the dead_philo boolean to decide
 // whether to continue the philosopher's routine or to terminate
 // the thread. It is protected behind the death mutex.
+// This function is also called in the print function to check
+// whether messages should stop being printed once a philosopher dies.
 bool	any_philo_dead(t_philo *philo)
 {
 	bool	dead;
